fix 64bit warnings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@4240 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ Texture.cpp
@@ -264,7 +264,7 @@ HRESULT Texture::SaveToStream(IStream *pstream, PinTable *pt)
 #if CURRENT_FILE_FORMAT_VERSION == 1060 // deprecated lower case name
    char tmp[sizeof(m_szName)];
    strncpy_s(tmp, m_szName, sizeof(tmp)-1);
-   const size_t len = strnlen_s(tmp, sizeof(tmp));
+   const DWORD len = (DWORD)strnlen_s(tmp, sizeof(tmp));
    CharLowerBuff(tmp, len);
    bw.WriteString(FID(INME), tmp);
 #endif
